requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
lxml
Flask
pageindex
python-dotenv
opik
pymupdf
EbookLib
html2text
